==> Templates/ModelProtoTypePipeline/ipynb/1_data_collection.ipynb <==
from package import *
import importlib
import utility as u
importlib.reload(u)
pass


conn = snowflake.connector.connect(
    user=u.SNOWFLAKE_USER,
    role=u.SNOWFLAKE_ROLE,
    private_key=u.SNOWFLAKE_PRIVATE_KEY,
    account="xxx",
    warehouse=u.SNOWFLAKE_WAREHOUSE
    )


data_collector = u.DataCollectorUtilClass(conn)
data_collector.run(u.START_DATE_TRAIN_DATA)
