==> Templates/ModelProtoTypePipeline/ipynb/3_data_cleaning.ipynb <==
from package import *
import importlib
import utility as u
importlib.reload(u)
pass


## Merge Data
* A document: a document represents the story of ...

df_1 = pd.read_parquet("../data/data_1.parquet")
df_2 = pd.read_parquet(f"../data/data_2.parquet")

data_util = u.DataUtilClass()
# using inner join led to a very small amount of rows in invoice_df dropped. Therefore, not worth it to use left join and impute missing value
df = data_util.merge(df_1, df_2, on = ["A"], how="inner")
print("using inner join led to retaining {:.2%} rows in invoice_df".format(len(df)/len(df_1)))

df.columns = [col.lower() for col in list(df.columns)]


# data in data.parquet include both paid and posted invoices
data_cleaner = u.DataCleanerUtilClass()
df = data_cleaner.run(df)
df.to_parquet(f"../data/data.parquet", index=False)
