==> Templates/ModelProtoTypePipeline/ipynb/4_training_data_prep.ipynb <==
from package import *
import importlib
import utility as u
importlib.reload(u)
pass


df = pd.read_parquet(f"../data/data.parquet")


### Feature Engineering

feature_engineer = u.FeatureEngineerUtilClass()
df_train = feature_engineer.run(df)


### Create Train Data

df_train.to_parquet(f"../data/data_train_test.parquet", index=False)
