==> Templates/ModelProtoTypePipeline/ipynb/5_EDA_train_data.ipynb <==
from package import *
import importlib
import utility as u
importlib.reload(u)
pass


eda_util = u.EDAUtilClass()

categorical_features = ["A", "B"]

historical_features = ["C", "D"]


df = pd.read_parquet(f"../data/data_train_test.parquet")
#df.profile_report()


# Marginal Analysis

## Distribution of Target


plt.figure(figsize=(8,6))
sns.histplot(df[u.TARGET_COL])
plt.title("abc")
plt.savefig(fname=f"../plot/EDA/target_distribution.png", bbox_inches="tight")
pass


# Generate x and y values for ECDF: x, y
plt.figure(figsize=(10,8))
x, y = dcst.ecdf(df["target"])
# Plot the ECDF as dots
_ = plt.plot(x, y, lw = 2)
_ = plt.xlabel('x', size = 14)
_ = plt.ylabel('ECDF', size = 14)
plt.xlim(0, df["target"].max())
plt.show()


## Distributions of Features

for feature in categorical_features:
    print(feature)
    display(eda_util.row_percentage_by_categorical_feature(df, feature))
    

plt.figure(figsize=(14, 8))

for i, feature in enumerate(historical_features):
    plt.subplot(2, 3, i+1)
    sns.histplot(df[feature])
    plt.title(f"histogram of {feature}")
plt.suptitle("x")
plt.tight_layout()
plt.savefig(fname=f"../plot/EDA/continuous_feature_distribution.png", bbox_inches="tight")


# Conditional Analysis

## Correlation Matrix

eda_util.correlation_heatmap(data=df[u.FEATURE_COLS+[u.TARGET_COL]], 
                             save_to_path=f"../plot/EDA/correlation_heatmap.png",
                             figsize=(14, 8))


## Variation of Target by Categorical Features

### Variation of Target by Feature A

df_tmp = df.groupby(["A"]).agg({"target":["mean","median"], 
                                "B":"nunique"
                               })
round(df_tmp, 0)


feature_col = "A"

eda_util.variation_of_target_by_feature(
    data=df,
    feature_col=feature_col, 
    target_col=u.TARGET_COL, 
    num_row=3, 
    num_col=2,
    figsize=(12,15),
    feature_type="categorical",
    target_type="categorical",
    sup_title=f"Distribution of Target Conditional on Different Values of {feature_col}",
    save_to_path=f"../plot/EDA/variation_of_target_distribution_by_feature_A.png"
)


## Variation of Target by Continuous Features

### Variation of Target by Feature B


for feature in historical_features:
    eda_util.variation_of_target_by_feature(
        data=df,
        feature_col=feature, 
        target_col=u.TARGET_COL, 
        num_row=1, 
        num_col=2,
        figsize=(12,5),
        feature_type="continuous",
        target_type="categorical",
        sup_title="x"
        save_to_path=f"../plot/EDA/variation_of_target_distribution_by_feature_B.png"
    )
